Converted from Vue component to a Vue Composable (Composition API function)
